+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix Holly Berry Leaves Drops
 - Fix Grass Generating in caves in Allium Shrubland
 - Adjust Just Like Grandma's Lang to say both pies
+- Make Stairs and Slabs Tags for BWG SandStone Sets and add to Respected Fabric Tags
 
 # 1.0.6-Beta
 - Fix Skyris Hanging Sign Textures being swapped